* move the custom hook to helpers.ts
