--- src/pages/private/Profile/Profile/Search/useDebounce.ts
@@ -1,15 +1,0 @@
-import { useEffect, useState } from 'react'
-
-export function useDebounce(value: any, delay: number) {
-    const [debouncedValue, setDebouncedValue] = useState(value)
-
-    useEffect(() => {
-        const handler = setTimeout(() => {
-            setDebouncedValue(value)
-        }, delay)
-
-        return () => clearTimeout(handler)
-    }, [value, delay])
-
-    return debouncedValue
-}
